+++ ErsatzTV.Core/Interfaces/Streaming/IGraphicsElement.cs
@@ -8,5 +8,10 @@ public interface IGraphicsElement
 
     Task InitializeAsync(Resolution frameSize, int frameRate, CancellationToken cancellationToken);
 
-    void Draw(object context, TimeSpan timestamp);
+    void Draw(
+        object context,
+        TimeSpan timeOfDay,
+        TimeSpan contentTime,
+        TimeSpan channelTime,
+        CancellationToken cancellationToken);
 }